Move range-proof-builder
